Fun with Widgets! You can now attach widgets to objects (event handlers not finished). Try these for some fun.
add_widgets(level.player, {type:"poly_line_widget",points:[[0,0],[0,10],[10,10],[10,0],[0,0]], color:[255,0,0,255], width:3})
add_widgets({type: "label", text: "Yawn!", size:14, x:0, y:0})
add_widgets({type:"button", label:{type:"label", text:"PushMe!"}, on_click:"debug('You Pressed Me!')"})

+++ src/property_editor_dialog.hpp
@@ -39,7 +39,7 @@ private:
 
 	typedef std::pair<gui::text_editor_widget_ptr, gui::slider_ptr> numeric_widgets;
 	void change_numeric_property(const std::string& id, boost::shared_ptr<numeric_widgets> w);
-	void change_numeric_property_slider(const std::string& id, boost::shared_ptr<numeric_widgets> w, float value);
+	void change_numeric_property_slider(const std::string& id, boost::shared_ptr<numeric_widgets> w, double value);
 	void change_enum_property(const std::string& id);
 	void set_enum_property(const std::string& id, const std::vector<std::string>& options, int index);
